refactor(orc8r): FegRelay servicers as External
